Fix tree
