+++ binary_ruby_tree.gemspec
@@ -1,13 +1,13 @@
 Gem::Specification.new do |s|
-  s.name = %q{bina_ruby_search_tree}
+  s.name = %q{binary_ruby_tree}
   s.version = "0.0.0"
   s.author = "Jonathan Fairbanks"
   s.date = %q{2019-07-30}
   s.summary = %q{Making BinaryTrees in ruby}
   s.files = [
-    "lib/bina_ruby_search_tree.rb",
-    "lib/bina_ruby_search_tree/node.rb",
-    "lib/bina_ruby_search_tree/binary_tree.rb",
+    "lib/binary_ruby_tree.rb",
+    "lib/binary_ruby_tree/node.rb",
+    "lib/binary_ruby_tree/tree.rb",
   ]
   s.require_paths = ["lib"]
 end
