+++ lib/binary_ruby_tree.rb
@@ -1,0 +1,10 @@
+require 'binary_ruby_tree/node'
+require 'binary_ruby_tree/tree'
+
+module BinaryRubyTree
+  module_function
+
+  def new_tree(array)
+    Tree.new(array)
+  end
+end
